Struct procure::cpu::CpuTimes

pub struct CpuTimes {
    pub user: u64,
    pub nice: u64,
    pub system: u64,
    pub idle: u64,
    pub iowait: u64,
    pub irq: u64,
    pub softirq: u64,
    pub steal: u64,
    pub guest: u64,
    pub guest_nice: u64,
}

Details about current CPU time utilization.

This object provides insight into the amount of time, measured in USER_HZ, that the system spent in various modes. These are hertz counters and not necessarily useful on their own, but against itself over some duration.

Fields

user

Time spent in user mode

nice

Time spent in user mode with low priority (nice)

system

Time spent in system mode

idle

Time spent idle

iowait

Time spent waiting for I/O to complete

irq

Time spent servicing interrupts

softirq

Time spent servicing softirqs

steal

Time spent in other operating systems when running in a virtualized environment.

guest

Time spent running a virtualized CPU for a guest operating system

guest_nice

Time spent running a niced guest virtualized CPU for a guest operating systems

Methods

impl CpuTimes

fn total() -> Result<CpuTimes>

Get a CpuTimes object filled with the total cpu times aggregated for all cores.

use procure::cpu::CpuTimes;

let times = CpuTimes::total().unwrap();
assert_eq!(
    times,
    CpuTimes {
        user: 7969864,
        nice: 6735,
        system: 1633028,
        iowait: 48613,
        idle: 43336958,
        irq: 180,
        softirq: 5043,
        steal: 0,
        guest: 0,
        guest_nice: 0,
    }
);

fn total_from_file(stat_file: &File) -> Result<CpuTimes>

Similar to CpuTimes::total but allows you to pass in an existing File to /proc/stat. This method expects the file cursor to be at 0.

fn per_cpu() -> Result<Vec<CpuTimes>>

Get a Vec of CpuTimes objects, one for each core.

use procure::cpu::CpuTimes;

let times = CpuTimes::per_cpu().unwrap();
assert_eq!(
    times,
    vec![
        CpuTimes {
            user: 2036657, nice: 3176, system: 538690, idle: 40502503, iowait: 48123,
            irq: 180, softirq: 4562, steal: 0, guest: 0, guest_nice: 0,
        },
        CpuTimes {
            user: 1895483, nice: 1224, system: 350858, idle: 947119, iowait: 194,
            irq: 0, softirq: 244, steal: 0, guest: 0, guest_nice: 0,
        },
    ]
);

fn per_cpu_from_file(stat_file: &File) -> Result<Vec<CpuTimes>>

Similar to CpuTimes::per_cpu but allows you to pass in an existing File to /proc/stat. This method expects the file cursor to be at 0.